refactor: move PortCheckerService out of gateway/ subpackage
PortCheckerService validates TCP ports (5555/5556) before they are
bound. It neither is a gateway nor talks to one; it was placed under
services/gateway/ because it shared the same boot sequence as the
market data gateway. Moving it up one level keeps services/gateway/
reserved for actual PFCF bridge components (DLL server/client, market
data gateway) and removes a cross-package surprise.

Updates CLAUDE.md path reference.
